Refactor Phase 5: extract render layers into three modules
Split 14 render functions from starfield.js into:
- render-background.js: Milky Way, twilight, grids, ecliptic, horizon
- render-objects.js: stars, DSOs, Sun, Moon, planets
- render-overlays.js: constellation lines/highlights, selection, labels

All render functions now take a shared render context object
{ ctx, cx, cy, scale, vf, fov } instead of reaching into module-level
state. Object renderers return screen buffers for hit testing.
starfield.js drops from ~866 to ~338 lines (app shell only).
