+++ js/starfield.js
@@ -1,34 +1,36 @@
 /**
- * starfield.js — Interactive Star Viewer
+ * starfield.js — Interactive Star Viewer (Application Shell)
  *
  * Real-time night sky from Los Angeles (34.05°N, 118.24°W).
  * Camera-frame projection in horizontal space — flat horizon at all azimuths.
  *
  * Architecture:
- *   js/sky/          — Pure astronomy math (time, projection, ephemeris). No DOM.
- *   js/viewer/       — Reusable modules: config, visual mapping, camera projection.
- *   js/starfield.js  — Application entry point: state, render loop, UI, interaction.
- *   css/starfield.css — All viewer UI styles (separate from portfolio styles).
+ *   js/sky/             — Pure astronomy math (time, projection, ephemeris). No DOM.
+ *   js/viewer/          — Feature modules: rendering, input, search, controls.
+ *   js/starfield.js     — Application shell: state, init, render loop.
+ *   css/starfield.css   — All viewer UI styles (separate from portfolio styles).
  *
  * Data: data/stars.json (generated by scripts/process_stars.py)
  */
 
 // --- Astronomy layer (pure math, no DOM) ---
 import { fovToScale }      from './sky/projection.js';
-import { lst, eqToHz }     from './sky/time.js';
+import { lst }             from './sky/time.js';
 import { planetPositions, sunPosition, moonPosition } from './sky/planets.js';
 
-// --- Viewer modules (reusable, no application state) ---
-import {
-  LAT_LA, LON_LA, D2R, R2D, FOV_DEFAULT, ALT_MIN, ALT_MAX,
-  BELOW_HORIZON_DIM, MAG_FADE_BAND, OBLIQUITY, CARDINALS, reducedMotion
-} from './viewer/config.js';
-import { bvToColor, magToRadius, magToAlpha, edgeFade, fovMagLimit } from './viewer/visual.js';
-import { buildViewFrame, projectStar, projectHzPoint } from './viewer/camera.js';
+// --- Viewer modules ---
+import { LON_LA, FOV_DEFAULT, ALT_MIN, ALT_MAX } from './viewer/config.js';
+import { fovMagLimit } from './viewer/visual.js';
+import { buildViewFrame } from './viewer/camera.js';
 import { advanceTime, getTimeState, setupTimeControls } from './viewer/controls.js';
 import { updatePopup } from './viewer/popup.js';
 import { buildSearchIndex, setupSearch } from './viewer/search.js';
 import { setupInput } from './viewer/input.js';
+
+// --- Render layers ---
+import { initMilkyWay, renderMilkyWay, renderTwilight, renderAltAzGrid, renderEqGrid, renderEcliptic, renderZodiacBand, renderHorizon } from './viewer/render-background.js';
+import { renderStars, renderDSOs, renderSun, renderMoon, renderPlanets } from './viewer/render-objects.js';
+import { renderConstellationLines, renderConstellationHighlight, renderSelection, renderCardinals, renderLabels } from './viewer/render-overlays.js';
 
 // --- DOM ---
 
@@ -46,36 +48,29 @@ let data = null;
 const view = { az: 180.0, alt: 45.0, fov: FOV_DEFAULT };
 const drag = { active: false, prevX: 0, prevY: 0, startX: 0, startY: 0, startTime: 0 };
 const overlays = { altAzGrid: false, eqGrid: false, ecliptic: true };
-const toggles  = { constellations: true }; // combined lines + labels
+const toggles  = { constellations: true };
 
-let selectedObject = null;  // { type, px, py, ra, dec, ... }
-let hoveredConst   = null;  // constellation abbr
-let clickedConst   = null;  // persisted constellation highlight
-let constFadeAlphas = {};    // { abbr: alpha } per-constellation fade values
+let selectedObject = null;
+let hoveredConst   = null;
+let clickedConst   = null;
+let constFadeAlphas = {};
 let _lastFrameTime = performance.now();
 
-// Time state managed by viewer/controls.js (advanceTime, getTimeState, togglePause, etc.)
-
-// Animated pan
-let viewTarget = null;      // { az, alt } or null — lerp target
-
-// Search managed by viewer/search.js
+let viewTarget = null;
 
 let W = 0, H = 0, cx = 0, cy = 0;
 
-// Screen position buffers for hit testing (filled each frame)
+// Screen position buffers for hit testing (filled each frame by render modules)
 let starScreenBuf = null;
 let starScreenCount = 0;
 let planetScreenBuf = [];
 let moonScreenPos = null;
 let constLabelScreen = [];
-let dsoScreenBuf = [];      // [{name, type, px, py, ra, dec, screenR}]
+let dsoScreenBuf = [];
 
-// Star name reverse lookup: "ra,dec" → {name, con, hipID}
+// Lookup tables (built at init)
 let starNameLookup = null;
-// HIP ID → constellation abbr (for hover-over-star → highlight constellation)
 let hipToConst = null;
-// Constellation abbr → constellation object (O(1) lookup)
 let constByAbbr = null;
 
 // --- Canvas sizing ---
@@ -93,539 +88,8 @@ function resize() {
   cy = H / 2;
 }
 
-// Formatters imported from viewer/popup.js
+// --- Popup delegation ---
 
-// Hit testing imported from viewer/hittest.js (pure functions with explicit parameters)
-
-// ============================================================
-// Render layers
-// ============================================================
-
-function renderStars(scale, vf) {
-  const cullCos = Math.cos((view.fov / 2 + 25) * D2R); // generous margin so stars at edges aren't clipped
-  const magLimit = fovMagLimit(view.fov);
-  const t = performance.now() * 0.001;
-
-  // Lazy-allocate screen buffer (px, py, starIndex per entry)
-  if (!starScreenBuf || starScreenBuf.length < data.stars.length * 3) {
-    starScreenBuf = new Float32Array(data.stars.length * 3);
-  }
-  starScreenCount = 0;
-
-  for (let i = 0; i < data.stars.length; i++) {
-    const s = data.stars[i];
-    const ra = s[0], dec = s[1], mag = s[2], ci = s[3];
-    if (mag > magLimit) break;
-
-    const p = projectStar(ra, dec, vf);
-    if (!p || p.cosAngle < cullCos) continue;
-
-    const px = cx + p.x * scale;
-    const py = cy - p.y * scale;
-    const r = magToRadius(mag);
-    let a = magToAlpha(mag) * (p.belowHorizon ? BELOW_HORIZON_DIM : 1.0);
-
-    if (mag > magLimit - MAG_FADE_BAND) a *= (magLimit - mag) / MAG_FADE_BAND;
-
-    if (mag < 4.0 && !p.belowHorizon && !reducedMotion) {
-      const phase = ((i * 2654435761) >>> 0) & 0xFFFF;
-      a *= 0.85 + 0.15 * (
-        0.6 * Math.sin(t * 3.7 + phase * 0.001) +
-        0.4 * Math.sin(t * 7.1 + phase * 0.0017)
-      );
-    }
-
-    const rgb = bvToColor(ci);
-    ctx.beginPath();
-    ctx.arc(px, py, r, 0, Math.PI * 2);
-    ctx.fillStyle = `rgba(${rgb},${a.toFixed(3)})`;
-    ctx.fill();
-
-    if (mag < 2.0 && !p.belowHorizon) {
-      const glowR = r * 5;
-      const grad = ctx.createRadialGradient(px, py, r * 0.5, px, py, glowR);
-      grad.addColorStop(0, `rgba(${rgb},${(a * 0.35).toFixed(3)})`);
-      grad.addColorStop(1, `rgba(${rgb},0)`);
-      ctx.beginPath();
-      ctx.arc(px, py, glowR, 0, Math.PI * 2);
-      ctx.fillStyle = grad;
-      ctx.fill();
-    }
-
-    // Write to screen buffer for hit testing (px, py, starIndex)
-    const bi = starScreenCount * 3;
-    starScreenBuf[bi] = px; starScreenBuf[bi+1] = py;
-    starScreenBuf[bi+2] = i; // original index into data.stars
-    starScreenCount++;
-  }
-}
-
-function renderConstellationLines(scale, vf) {
-  const cullCos = Math.cos((view.fov / 2 + 30) * D2R);
-
-  for (const c of data.constellations) {
-    const ha = constFadeAlphas[c.abbr] || 0;
-    if (!toggles.constellations && ha <= 0) continue;
-
-    if (ha > 0) {
-      const a = 0.35 + (0.65 - 0.35) * ha;
-      ctx.strokeStyle = `rgba(${Math.round(255 - 155*ha)},${Math.round(255 - 75*ha)},255,${a.toFixed(3)})`;
-      ctx.lineWidth = 1.2 + 1.0 * ha;
-    } else {
-      ctx.strokeStyle = 'rgba(255,255,255,0.35)';
-      ctx.lineWidth = 1.2;
-    }
-
-    for (const seg of c.resolvedLines) {
-      const p1 = projectStar(seg[0], seg[1], vf);
-      const p2 = projectStar(seg[2], seg[3], vf);
-      if (!p1 || !p2) continue;
-      if (p1.cosAngle < cullCos && p2.cosAngle < cullCos) continue;
-      ctx.beginPath();
-      ctx.moveTo(cx + p1.x * scale, cy - p1.y * scale);
-      ctx.lineTo(cx + p2.x * scale, cy - p2.y * scale);
-      ctx.stroke();
-    }
-  }
-}
-
-function renderConstellationHighlight(scale, vf) {
-  // Draw highlight rings for ALL constellations that have a non-zero fade alpha
-  for (const abbr in constFadeAlphas) {
-    const ha = constFadeAlphas[abbr];
-    if (ha <= 0.01) continue;
-    const con = constByAbbr.get(abbr);
-    if (!con) continue;
-
-    const seen = new Set();
-    ctx.strokeStyle = `rgba(100,180,255,${(0.8 * ha).toFixed(3)})`;
-    ctx.lineWidth = 1.5;
-
-    for (const [h1, h2] of con.lines) {
-      for (const h of [h1, h2]) {
-        if (seen.has(h)) continue;
-        seen.add(h);
-        const pos = data.hip[String(h)];
-        if (!pos) continue;
-        const p = projectStar(pos[0], pos[1], vf);
-        if (!p) continue;
-        ctx.beginPath();
-        ctx.arc(cx + p.x * scale, cy - p.y * scale, 6, 0, Math.PI * 2);
-        ctx.stroke();
-      }
-    }
-  }
-}
-
-function renderSelection(scale, vf) {
-  if (!selectedObject) return;
-  const p = projectStar(selectedObject.ra, selectedObject.dec, vf);
-  if (!p) return;
-  const px = cx + p.x * scale, py = cy - p.y * scale;
-  const pulse = reducedMotion ? 1.0 : 0.7 + 0.3 * Math.sin(performance.now() * 0.003);
-
-  ctx.strokeStyle = `rgba(255,220,100,${pulse.toFixed(2)})`;
-  ctx.lineWidth = 2;
-  ctx.beginPath();
-  ctx.arc(px, py, 8, 0, Math.PI * 2);
-  ctx.stroke();
-
-  // Update popup position
-  if (popupEl) {
-    selectedObject.px = px;
-    selectedObject.py = py;
-  }
-}
-
-function renderDSOs(scale, vf) {
-  const cullCos = Math.cos((view.fov / 2 + 25) * D2R);
-  dsoScreenBuf = [];
-
-  for (let di = 0; di < data.dsos.length; di++) {
-    const d = data.dsos[di];
-    const p = projectStar(d.ra, d.dec, vf);
-    if (!p || p.cosAngle < cullCos) continue;
-
-    const px = cx + p.x * scale, py = cy - p.y * scale;
-    const screenR = Math.max(6, d.size * D2R * scale);
-    const b = d.brightness;
-
-    if (d.type === 'nebula') {
-      ctx.fillStyle = `rgba(255,180,200,${(b * 0.25).toFixed(3)})`;
-      ctx.beginPath(); ctx.arc(px, py, screenR * 0.4, 0, Math.PI * 2); ctx.fill();
-      ctx.fillStyle = `rgba(200,150,180,${(b * 0.10).toFixed(3)})`;
-      ctx.beginPath(); ctx.arc(px, py, screenR, 0, Math.PI * 2); ctx.fill();
-    } else if (d.type === 'galaxy') {
-      const angle = (di * 137.508) * D2R;
-      ctx.save(); ctx.translate(px, py); ctx.rotate(angle);
-      ctx.fillStyle = `rgba(255,240,200,${(b * 0.20).toFixed(3)})`;
-      ctx.beginPath(); ctx.ellipse(0, 0, screenR * 0.6, screenR * 0.25, 0, 0, Math.PI * 2); ctx.fill();
-      ctx.fillStyle = `rgba(255,230,180,${(b * 0.08).toFixed(3)})`;
-      ctx.beginPath(); ctx.ellipse(0, 0, screenR, screenR * 0.45, 0, 0, Math.PI * 2); ctx.fill();
-      ctx.restore();
-    } else if (d.type === 'open_cluster') {
-      ctx.fillStyle = `rgba(200,220,255,${(b * 0.35).toFixed(3)})`;
-      for (let j = 0; j < 7; j++) {
-        const a2 = (j * 137.508 + di * 50) * D2R;
-        const r2 = screenR * 0.4 * (0.3 + (j % 3) * 0.35);
-        ctx.beginPath(); ctx.arc(px + Math.cos(a2) * r2, py + Math.sin(a2) * r2, 1.5, 0, Math.PI * 2); ctx.fill();
-      }
-    } else if (d.type === 'globular_cluster') {
-      ctx.fillStyle = `rgba(255,240,210,${(b * 0.35).toFixed(3)})`;
-      ctx.beginPath(); ctx.arc(px, py, screenR * 0.25, 0, Math.PI * 2); ctx.fill();
-      ctx.fillStyle = `rgba(255,230,200,${(b * 0.12).toFixed(3)})`;
-      ctx.beginPath(); ctx.arc(px, py, screenR * 0.6, 0, Math.PI * 2); ctx.fill();
-      ctx.fillStyle = `rgba(240,220,190,${(b * 0.05).toFixed(3)})`;
-      ctx.beginPath(); ctx.arc(px, py, screenR, 0, Math.PI * 2); ctx.fill();
-    }
-
-    // Selection highlight for DSOs
-    if (selectedObject && selectedObject.type === 'dso' && selectedObject.name === d.name) {
-      const pulse = reducedMotion ? 1.0 : 0.7 + 0.3 * Math.sin(performance.now() * 0.003);
-      ctx.strokeStyle = `rgba(180,180,255,${pulse.toFixed(2)})`;
-      ctx.lineWidth = 1.5;
-      ctx.beginPath(); ctx.arc(px, py, screenR + 4, 0, Math.PI * 2); ctx.stroke();
-    }
-
-    dsoScreenBuf.push({ name: d.name, type: d.type, px, py, ra: d.ra, dec: d.dec, screenR });
-  }
-}
-
-// --- Milky Way ---
-
-let milkyWayPoints = null;
-
-function buildMilkyWayPoints(waypoints) {
-  const wp = (waypoints[0].ra === waypoints[waypoints.length - 1].ra &&
-              waypoints[0].dec === waypoints[waypoints.length - 1].dec)
-    ? waypoints.slice(0, -1) : waypoints;
-  const pts = [], n = wp.length, STEPS = 8;
-  for (let i = 0; i < n; i++) {
-    const a = wp[(i-1+n)%n], b = wp[i], c = wp[(i+1)%n], d = wp[(i+2)%n];
-    for (let s = 0; s < STEPS; s++) {
-      const t = s / STEPS, t2 = t*t, t3 = t2*t;
-      const ra  = 0.5*((2*b.ra)+(-a.ra+c.ra)*t+(2*a.ra-5*b.ra+4*c.ra-d.ra)*t2+(-a.ra+3*b.ra-3*c.ra+d.ra)*t3);
-      const dec = 0.5*((2*b.dec)+(-a.dec+c.dec)*t+(2*a.dec-5*b.dec+4*c.dec-d.dec)*t2+(-a.dec+3*b.dec-3*c.dec+d.dec)*t3);
-      const w = b.width + (c.width - b.width) * t;
-      const seed = ((i*8+s)*2654435761)>>>0;
-      const rA = ((seed&0xFFFF)/0xFFFF-0.5), rD = (((seed>>>16)&0xFFFF)/0xFFFF-0.5);
-      pts.push({ra, dec, width: w*0.5});
-      pts.push({ra: ra+rA*w*0.02, dec: dec+rD*w*0.3, width: w*0.3});
-    }
-  }
-  return pts;
-}
-
-function renderMilkyWay(scale, vf) {
-  if (!milkyWayPoints) return;
-  ctx.fillStyle = 'rgba(180,180,210,0.008)';
-  for (const pt of milkyWayPoints) {
-    const p = projectStar(pt.ra, pt.dec, vf);
-    if (!p || p.cosAngle < Math.cos((view.fov/2+pt.width)*D2R)) continue;
-    const px = cx + p.x * scale, py = cy - p.y * scale;
-    ctx.beginPath(); ctx.arc(px, py, pt.width*D2R*scale, 0, Math.PI*2); ctx.fill();
-  }
-}
-
-// --- Twilight glow ---
-
-function renderTwilight(scale, vf, lstDeg) {
-  const sunHz = eqToHz(_cachedSun.ra, _cachedSun.dec, lstDeg, LAT_LA);
-
-  if (sunHz.alt > 0 || sunHz.alt < -18) return;
-
-  const horizonP = projectHzPoint(0, sunHz.az, vf);
-  if (!horizonP) return;
-
-  const px = cx + horizonP.x * scale, py = cy - horizonP.y * scale;
-  const depth = -sunHz.alt;
-  const intensity = Math.max(0, 1 - depth / 18);
-  const radius = 200 * intensity + 50;
-
-  const r = depth < 6 ? 255 : Math.round(255 * Math.max(0, 1 - (depth-6)/12));
-  const g = depth < 6 ? Math.round(180 - depth * 15) : Math.round(100 * Math.max(0, 1 - (depth-6)/12));
-  const b = depth < 6 ? Math.round(100 + depth * 10) : Math.round(160 * Math.max(0, 1 - (depth-6)/12));
-
-  const grad = ctx.createRadialGradient(px, py, 0, px, py, radius);
-  grad.addColorStop(0, `rgba(${r},${g},${b},${(intensity * 0.3).toFixed(3)})`);
-  grad.addColorStop(1, 'rgba(0,0,0,0)');
-  ctx.fillStyle = grad;
-  ctx.beginPath(); ctx.arc(px, py, radius, 0, Math.PI*2); ctx.fill();
-}
-
-// --- Sun ---
-
-function renderSun(scale, vf) {
-  const sun = _cachedSun;
-  if (!sun) return;
-  const p = projectStar(sun.ra, sun.dec, vf);
-  if (!p) return;
-  const alpha = edgeFade(p.cosAngle, view.fov);
-  if (alpha <= 0) return;
-
-  const px = cx + p.x * scale, py = cy - p.y * scale;
-  const sunR = 10; 
-
-  // Yellow core
-  ctx.beginPath();
-  ctx.arc(px, py, sunR, 0, Math.PI * 2);
-  ctx.fillStyle = `rgba(255,220,50,${(alpha * 0.95).toFixed(3)})`;
-  ctx.fill();
-
-  // Warm glow
-  const grad = ctx.createRadialGradient(px, py, sunR, px, py, sunR * 4);
-  grad.addColorStop(0, `rgba(255,200,50,${(alpha * 0.3).toFixed(3)})`);
-  grad.addColorStop(1, 'rgba(255,180,30,0)');
-  ctx.beginPath();
-  ctx.arc(px, py, sunR * 4, 0, Math.PI * 2);
-  ctx.fillStyle = grad;
-  ctx.fill();
-
-  // Label
-  ctx.save();
-  ctx.textAlign = 'center';
-  ctx.textBaseline = 'top';
-  ctx.font = '11px sans-serif';
-  ctx.fillStyle = `rgba(255,220,100,${(alpha * 0.85).toFixed(3)})`;
-  ctx.fillText('Sun', px, py + sunR + 4);
-  ctx.restore();
-}
-
-// --- Moon ---
-
-function renderMoon(scale, vf) {
-  const moon = _cachedMoon;
-  const p = projectStar(moon.ra, moon.dec, vf);
-  if (!p) { moonScreenPos = null; return; }
-
-  const alpha = edgeFade(p.cosAngle, view.fov);
-  if (alpha <= 0) { moonScreenPos = null; return; }
-
-  const px = cx + p.x * scale, py = cy - p.y * scale;
-  const moonR = 5;
-
-  // Illuminated fraction and phase side
-  const k = (1 + Math.cos(moon.elongation * D2R)) / 2;
-  // Determine waxing vs waning: moon east of sun = waxing (shadow on right)
-  const sunPos = _cachedSun;
-  const eastOfSun = ((sunPos.ra - moon.ra) * 15 + 360) % 360 > 180;
-  const shadowSign = eastOfSun ? 1 : -1;
-
-  // Draw bright disc
-  ctx.beginPath(); ctx.arc(px, py, moonR, 0, Math.PI*2);
-  ctx.fillStyle = `rgba(230,230,220,${(alpha * 0.9).toFixed(3)})`;
-  ctx.fill();
-
-  // Shadow overlay for phase
-  if (k < 0.98) {
-    ctx.save();
-    ctx.beginPath(); ctx.arc(px, py, moonR, 0, Math.PI*2); ctx.clip();
-    const shadowX = px + moonR * (2 * k - 1) * shadowSign;
-    ctx.fillStyle = `rgba(5,5,16,${(alpha * 0.85).toFixed(3)})`;
-    ctx.beginPath(); ctx.arc(shadowX, py, moonR, 0, Math.PI*2); ctx.fill();
-    ctx.restore();
-  }
-
-  // Soft glow
-  ctx.beginPath(); ctx.arc(px, py, moonR * 3, 0, Math.PI*2);
-  const grad = ctx.createRadialGradient(px, py, moonR, px, py, moonR*3);
-  grad.addColorStop(0, `rgba(200,200,190,${(alpha*0.15).toFixed(3)})`);
-  grad.addColorStop(1, 'rgba(200,200,190,0)');
-  ctx.fillStyle = grad; ctx.fill();
-
-  // Label
-  ctx.save();
-  ctx.textAlign = 'center'; ctx.textBaseline = 'top';
-  ctx.font = '11px sans-serif';
-  ctx.fillStyle = `rgba(220,220,210,${(alpha*0.8).toFixed(3)})`;
-  ctx.fillText('Moon', px, py + moonR + 4);
-  ctx.restore();
-
-  // Store for hit testing
-  moonScreenPos = { name: 'Moon', px, py, ra: moon.ra, dec: moon.dec,
-                    elongation: moon.elongation, illumination: k, screenR: moonR };
-}
-
-// --- Grid overlays ---
-
-function drawGridLine(count, getPoint, scale) {
-  let penDown = false;
-  for (let i = 0; i < count; i++) {
-    const p = getPoint(i);
-    if (!p || p.cosAngle < 0.01) { penDown = false; continue; }
-    const px = cx + p.x * scale, py = cy - p.y * scale;
-    if (!penDown) { ctx.moveTo(px, py); penDown = true; }
-    else ctx.lineTo(px, py);
-  }
-}
-
-function renderAltAzGrid(scale, vf) {
-  ctx.strokeStyle = 'rgba(100,150,200,0.25)'; ctx.lineWidth = 0.5;
-  for (let alt = -80; alt <= 80; alt += 10) {
-    if (alt === 0) continue;
-    ctx.beginPath(); drawGridLine(181, i => projectHzPoint(alt, i*2, vf), scale); ctx.stroke();
-  }
-  for (let az = 0; az < 360; az += 15) {
-    ctx.beginPath(); drawGridLine(91, i => projectHzPoint(-90+i*2, az, vf), scale); ctx.stroke();
-  }
-}
-
-function renderEqGrid(scale, vf) {
-  for (let dec = -80; dec <= 80; dec += 10) {
-    ctx.strokeStyle = dec === 0 ? 'rgba(100,200,150,0.5)' : 'rgba(100,200,150,0.25)';
-    ctx.lineWidth = dec === 0 ? 1.0 : 0.5;
-    ctx.beginPath(); drawGridLine(181, i => projectStar(i*24/180, dec, vf), scale); ctx.stroke();
-  }
-  ctx.strokeStyle = 'rgba(100,200,150,0.25)'; ctx.lineWidth = 0.5;
-  for (let raH = 0; raH < 24; raH++) {
-    ctx.beginPath(); drawGridLine(91, i => projectStar(raH, -90+i*2, vf), scale); ctx.stroke();
-  }
-}
-
-function renderEcliptic(scale, vf) {
-  const oblR = OBLIQUITY * D2R;
-  ctx.strokeStyle = 'rgba(200,200,100,0.35)'; ctx.lineWidth = 1.0;
-  ctx.setLineDash([6, 4]);
-  ctx.beginPath();
-  drawGridLine(361, i => {
-    const lamR = i * D2R;
-    return projectStar(((Math.atan2(Math.sin(lamR)*Math.cos(oblR), Math.cos(lamR))*R2D+360)%360)/15,
-                       Math.asin(Math.sin(lamR)*Math.sin(oblR))*R2D, vf);
-  }, scale);
-  ctx.stroke(); ctx.setLineDash([]);
-}
-
-/** Zodiac band — ±9° around the ecliptic where constellations and planets live. */
-function renderZodiacBand(scale, vf) {
-  const oblR = OBLIQUITY * D2R;
-  const BAND_HALF = 9; // degrees above/below ecliptic
-
-  // Draw two parallel lines at ecliptic latitude ±BAND_HALF
-  for (const offset of [BAND_HALF, -BAND_HALF]) {
-    ctx.strokeStyle = 'rgba(40,120,200,0.18)';
-    ctx.lineWidth = 0.8;
-    ctx.beginPath();
-    drawGridLine(361, i => {
-      const lamR = i * D2R;
-      // Ecliptic longitude → ecliptic lat offset → equatorial RA/Dec
-      const eclLat = offset * D2R;
-      const sinLam = Math.sin(lamR), cosLam = Math.cos(lamR);
-      const sinB = Math.sin(eclLat), cosB = Math.cos(eclLat);
-      // Ecliptic → equatorial rotation
-      const x = cosB * cosLam;
-      const y = cosB * sinLam * Math.cos(oblR) - sinB * Math.sin(oblR);
-      const z = cosB * sinLam * Math.sin(oblR) + sinB * Math.cos(oblR);
-      const ra = ((Math.atan2(y, x) * R2D + 360) % 360) / 15;
-      const dec = Math.asin(Math.max(-1, Math.min(1, z))) * R2D;
-      return projectStar(ra, dec, vf);
-    }, scale);
-    ctx.stroke();
-  }
-}
-
-// --- Horizon, Cardinals ---
-
-function renderHorizon(scale, vf) {
-  ctx.beginPath(); let penDown = false;
-  for (let i = 0; i <= 360; i++) {
-    const p = projectHzPoint(0, i%360, vf);
-    if (!p || p.cosAngle < 0.01) { penDown = false; continue; }
-    const px = cx+p.x*scale, py = cy-p.y*scale;
-    if (!penDown) { ctx.moveTo(px, py); penDown = true; } else ctx.lineTo(px, py);
-  }
-  ctx.strokeStyle = 'rgba(255,255,255,0.55)'; ctx.lineWidth = 1.5; ctx.stroke();
-}
-
-function renderPlanets(scale, vf) {
-  const cullCos = Math.cos((view.fov / 2 + 25) * D2R);
-  const planets = _cachedPlanets;
-  planetScreenBuf = [];
-
-  for (const planet of planets) {
-    const p = projectStar(planet.ra, planet.dec, vf);
-    if (!p || p.cosAngle < cullCos) continue;
-    const alpha = edgeFade(p.cosAngle, view.fov);
-    if (alpha <= 0) continue;
-
-    const a = alpha * (p.belowHorizon ? BELOW_HORIZON_DIM : 1.0);
-    const px = cx + p.x * scale, py = cy - p.y * scale;
-    const r = planet.screenR;
-
-    const grad = ctx.createRadialGradient(px, py, 0, px, py, r*2.5);
-    grad.addColorStop(0, `rgba(255,165,0,${(a*0.9).toFixed(3)})`);
-    grad.addColorStop(0.5, `rgba(255,140,0,${(a*0.4).toFixed(3)})`);
-    grad.addColorStop(1, 'rgba(255,120,0,0)');
-    ctx.beginPath(); ctx.arc(px, py, r*2.5, 0, Math.PI*2); ctx.fillStyle = grad; ctx.fill();
-    ctx.beginPath(); ctx.arc(px, py, r, 0, Math.PI*2);
-    ctx.fillStyle = `rgba(255,180,60,${a.toFixed(3)})`; ctx.fill();
-
-    ctx.save();
-    ctx.textAlign = 'center'; ctx.textBaseline = 'top'; ctx.font = '11px sans-serif';
-    ctx.fillStyle = `rgba(255,200,100,${(a*0.85).toFixed(3)})`;
-    ctx.fillText(planet.name, px, py + r + 4);
-    ctx.restore();
-
-    planetScreenBuf.push({ name: planet.name, px, py, ra: planet.ra, dec: planet.dec, screenR: r });
-  }
-}
-
-function renderCardinals(scale, vf) {
-  ctx.font = 'bold 13px monospace'; ctx.textAlign = 'center'; ctx.textBaseline = 'bottom';
-  for (const { az, label } of CARDINALS) {
-    const p = projectHzPoint(0, az, vf);
-    if (!p || p.cosAngle < 0.01) continue;
-    const alpha = edgeFade(p.cosAngle, view.fov);
-    if (alpha <= 0) continue;
-    const px = cx+p.x*scale, py = cy-p.y*scale;
-    ctx.strokeStyle = `rgba(255,255,255,${(0.75*alpha).toFixed(3)})`; ctx.lineWidth = 1.5;
-    ctx.beginPath(); ctx.moveTo(px, py-7); ctx.lineTo(px, py+7); ctx.stroke();
-    ctx.fillStyle = `rgba(255,255,255,${(0.9*alpha).toFixed(3)})`;
-    ctx.fillText(label, px, py-10);
-  }
-}
-
-function renderLabels(scale, vf) {
-  constLabelScreen = [];
-
-  if (toggles.constellations) {
-    ctx.textAlign = 'center'; ctx.textBaseline = 'bottom'; ctx.font = '13px sans-serif';
-    for (const c of data.constellations) {
-      const p = projectStar(c.label_ra, c.label_dec, vf);
-      if (!p) continue;
-      const alpha = edgeFade(p.cosAngle, view.fov);
-      if (alpha <= 0) continue;
-      const px = cx + p.x * scale, py = cy - p.y * scale;
-      const ha = constFadeAlphas[c.abbr] || 0;
-      const r = Math.round(255 - 155 * ha), g = Math.round(255 - 75 * ha);
-      const a2 = (0.45 + 0.40 * ha) * alpha;
-      ctx.fillStyle = `rgba(${r},${g},255,${a2.toFixed(3)})`;
-      ctx.fillText(c.name, px, py - 10);
-      constLabelScreen.push({ abbr: c.abbr, px, py });
-    }
-  } else {
-    // Still fill constLabelScreen for hover detection even when labels hidden
-    for (const c of data.constellations) {
-      const p = projectStar(c.label_ra, c.label_dec, vf);
-      if (!p) continue;
-      if (edgeFade(p.cosAngle, view.fov) <= 0) continue;
-      constLabelScreen.push({ abbr: c.abbr, px: cx + p.x * scale, py: cy - p.y * scale });
-    }
-  }
-
-  ctx.textBaseline = 'top'; ctx.font = '11px sans-serif';
-  for (const d of data.dsos) {
-    const p = projectStar(d.ra, d.dec, vf);
-    if (!p) continue;
-    const alpha = edgeFade(p.cosAngle, view.fov);
-    if (alpha <= 0) continue;
-    ctx.fillStyle = `rgba(180,180,255,${(0.40*alpha).toFixed(3)})`;
-    ctx.fillText(d.name, cx+p.x*scale, cy-p.y*scale+10);
-  }
-}
-
-// Info popup imported from viewer/popup.js
-
-// updatePopup delegates to viewer/popup.js with application state as dependencies
 function callUpdatePopup() {
   updatePopup(popupEl, selectedObject, clickedConst, {
     cachedPlanets: _cachedPlanets, cachedMoon: _cachedMoon,
@@ -683,7 +147,7 @@ function updateInfo() {
     g(overlays.ecliptic, 'E') + ' ' + g(toggles.constellations, 'C');
 }
 
-// --- Ephemeris cache (planets/sun/moon recomputed once per minute, not per frame) ---
+// --- Ephemeris cache ---
 
 let _cachedPlanets = null;
 let _cachedSun = null;
@@ -691,7 +155,7 @@ let _cachedMoon = null;
 let _lastEphemJD = 0;
 
 function updateEphemeris(jd) {
-  if (jd - _lastEphemJD > 1 / 1440) { // 1 minute in Julian days
+  if (jd - _lastEphemJD > 1 / 1440) {
     _cachedPlanets = planetPositions(jd);
     _cachedSun = sunPosition(jd);
     _cachedMoon = moonPosition(jd);
@@ -702,13 +166,11 @@ function updateEphemeris(jd) {
 // --- Main render ---
 
 function render() {
-  // Advance simulated time (managed by viewer/controls.js)
   const jd = advanceTime();
-
   const lstDeg = lst(jd, LON_LA);
   updateEphemeris(jd);
 
-  // Animate per-frame timing
+  // Per-frame timing
   const now = performance.now();
   const dt = Math.min((now - _lastFrameTime) / 1000, 0.1);
   _lastFrameTime = now;
@@ -727,11 +189,14 @@ function render() {
     }
   }
 
+  // Build camera frame and render context
   const vf = buildViewFrame(view.alt, view.az, lstDeg);
   const scale = fovToScale(view.fov, Math.min(cx, cy));
+  const rc = { ctx, cx, cy, scale, vf, fov: view.fov };
+
+  // Constellation fade animation
   const fadeSpeed = dt / 0.3;
   const activeAbbr = hoveredConst || clickedConst;
-  // Fade in active constellation, fade out all others
   for (const abbr in constFadeAlphas) {
     if (abbr === activeAbbr) continue;
     constFadeAlphas[abbr] -= fadeSpeed;
@@ -741,25 +206,36 @@ function render() {
     constFadeAlphas[activeAbbr] = Math.min(1, (constFadeAlphas[activeAbbr] || 0) + fadeSpeed);
   }
 
+  // Lazy-allocate star screen buffer
+  if (!starScreenBuf || starScreenBuf.length < data.stars.length * 3) {
+    starScreenBuf = new Float32Array(data.stars.length * 3);
+  }
+
+  // Clear
   ctx.fillStyle = '#050510';
   ctx.fillRect(0, 0, W, H);
 
-  renderMilkyWay(scale, vf);
-  renderTwilight(scale, vf, lstDeg);
-  if (overlays.altAzGrid) renderAltAzGrid(scale, vf);
-  if (overlays.eqGrid) renderEqGrid(scale, vf);
-  if (overlays.ecliptic) { renderZodiacBand(scale, vf); renderEcliptic(scale, vf); }
-  renderDSOs(scale, vf);
-  renderConstellationLines(scale, vf);
-  renderStars(scale, vf);
-  renderConstellationHighlight(scale, vf);
-  renderSelection(scale, vf);
-  renderHorizon(scale, vf);
-  renderCardinals(scale, vf);
-  renderPlanets(scale, vf);
-  renderSun(scale, vf);
-  renderMoon(scale, vf);
-  renderLabels(scale, vf);
+  // Background layers
+  renderMilkyWay(rc);
+  renderTwilight(rc, lstDeg, _cachedSun);
+  if (overlays.altAzGrid) renderAltAzGrid(rc);
+  if (overlays.eqGrid) renderEqGrid(rc);
+  if (overlays.ecliptic) { renderZodiacBand(rc); renderEcliptic(rc); }
+
+  // Sky objects (return screen buffers for hit testing)
+  dsoScreenBuf = renderDSOs(rc, data.dsos, selectedObject);
+  renderConstellationLines(rc, data.constellations, constFadeAlphas, toggles.constellations);
+  starScreenCount = renderStars(rc, data.stars, starScreenBuf);
+  renderConstellationHighlight(rc, constFadeAlphas, constByAbbr, data.hip);
+  renderSelection(rc, selectedObject);
+  renderHorizon(rc);
+  renderCardinals(rc);
+  planetScreenBuf = renderPlanets(rc, _cachedPlanets);
+  renderSun(rc, _cachedSun);
+  moonScreenPos = renderMoon(rc, _cachedMoon, _cachedSun);
+  constLabelScreen = renderLabels(rc, data.constellations, data.dsos, constFadeAlphas, toggles.constellations);
+
+  // UI updates
   updateInfo();
   updateClock();
   callUpdatePopup();
@@ -769,8 +245,6 @@ function frame() {
   render();
   requestAnimationFrame(frame);
 }
-
-// Input, search, and time controls managed by viewer modules
 
 // --- Init ---
 
@@ -784,7 +258,6 @@ function buildStarNameLookup() {
       starNameLookup.set(key, { ...data.star_names[hipID], hipID });
     }
   }
-  // Build ra,dec → constellation abbr for all constellation member stars
   hipToConst = new Map();
   for (const c of data.constellations) {
     for (const [h1, h2] of c.lines) {
@@ -820,11 +293,10 @@ async function init() {
     }
   }
 
-  milkyWayPoints = buildMilkyWayPoints(data.milky_way);
+  initMilkyWay(data.milky_way);
   buildStarNameLookup();
   constByAbbr = new Map(data.constellations.map(c => [c.abbr, c]));
 
-  // Initialize ephemeris so search index has planet positions
   const initJd = (Date.now() + getTimeState().timeOffsetMs) / 86400000 + 2440587.5;
   updateEphemeris(initJd);
   buildSearchIndex(data, _cachedPlanets, _cachedMoon);
